feature(fronted): toaster && confirmDelete users

+++ src/app/users/page.js
@@ -7,7 +7,7 @@ import Edit from "@/components/Edit";
 import Button from "@/components/UI/Button";
 import CardUser from "@/components/UI/card/user_card";
 import FormUser from "@/components/UI/form/user_form";
-import { Toaster } from 'react-hot-toast';
+import { Toaster, toast } from 'react-hot-toast';
 
 export default function Users() {
     const title = "users";
@@ -40,11 +40,21 @@ export default function Users() {
     }, [db_name]);
 
     const deleteData = async (id) => {
-        try {
-            await FetchApi({ url: `/api/${db_name}/${id}`, method: "DELETE" });
-            setDataList(dataList.filter((item) => item.id !== id));
-        } catch (error) {
-            console.error("Deletion error: ", error);
+        const confirmDelete = window.confirm(
+            "Êtes-vous sûr de vouloir supprimer cet élément ?"
+        );
+        if (confirmDelete) {
+            try {
+                await FetchApi({ url: `/api/${db_name}/${id}`, method: "DELETE" });
+                setDataList(dataList.filter((item) => item.id !== id));
+                console.log(`User with ID ${id} deleted successfully`);
+                toast.success(`User deleted successfully`);
+            } 
+            catch (error) {
+                console.error("Deletion error: ", error);
+                console.log(`Failed to delete user with ID ${id}`);
+                toast.error(`Failed to delete user with ID ${id}`);
+            }   
         }
     };
